requirements.txt
# 核心框架
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 浏览器自动化
playwright==1.40.0

# 验证码处理
pytesseract==0.3.10
opencv-python==4.8.1.5
Pillow==10.1.0

# HTTP 客户端
requests==2.31.0
aiohttp==3.9.1

# 工具库
python-dotenv==1.0.0
loguru==0.7.2
python-json-logger==2.0.7

# 监控和日志
prometheus-client==0.19.0
psutil==5.9.6

# 可选：LLM 集成
openai==1.3.8

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# 开发工具
black==23.12.0
flake8==6.1.0
mypy==1.7.1